Fixed two warnings
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@5326 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/items/powerup_manager.cpp
@@ -198,7 +198,7 @@ void PowerupManager::loadWeights(const XMLNode &root,
         printf("Incorrect number of weights found in class '%s':\n", 
                class_name.c_str());
         printf("%d instead of %d - probabilities will be incorrect.\n",
-               weight_list.size(), (int)POWERUP_LAST);
+               (int)weight_list.size(), (int)POWERUP_LAST);
         return;
     }
